% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_emissions_epci}
\alias{data_emissions_epci}
\title{Emissions information for EPCI}
\format{An object of class \code{data.frame} with 53 rows and 16 columns.}
\usage{
data_emissions_epci
}
\description{
emissions information for epci stored in three
data.frame. The dataset contains the
following columns:
\itemize{
\item \code{gid}: group identifier
\item \code{lib_epci}: public intermunicipal cooperation institute name
\item \code{annee_inv}: year
\item \code{pm10}: Particulate matter with a diameter of 10 micrometers or less (in kg)
\item \code{pm25}: Particulate matter with a diameter of 25 micrometers or less (in kg)
\item \code{nox}: Nitrogen oxide (in kg)
\item \code{so2}: Sulfur dioxide (in kg)
\item \code{co}: Carbon monoxide (in kg)
\item \code{nh3}: Ammonia (in kg)
\item \code{covnm}: Non-methane volatile organic compounds (in kg)
\item \code{as}: Arsenic (in kg)
\item \code{cd}: Cadmium (in kg)
\item \code{ni}: Nickel (in kg)
\item \code{pb}: Lead (in kg)
\item \code{bap}: Benzo(a)pyrene (in kg)
\item \code{co2_t}: Carbon dioxide. Measure for this emission is tons.
}
}
\references{
\url{https://data-airparif-asso.opendata.arcgis.com}
}
\author{
Airparif
}
\keyword{data}
