% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emissions_region}
\alias{emissions_region}
\title{Pollutants emissions for 2015 by region}
\format{An object of class \code{data.frame} with 1 rows and 12 columns.}
\usage{
emissions_region
}
\description{
Regional pollutants emissions
as a data.frame. The dataset contains the following columns:
\itemize{
\item \code{gid}: geographical id
\item \code{lib_reg}: region label
\item \code{code_reg}: region code
\item \code{year}: year of emissions
\item \code{pm10}: Particulate matter with a diameter of 10 micrometers or less (in kg)
\item \code{pm25}: Particulate matter with a diameter of 2.5 micrometers or less (in kg)
\item \code{nox}: Nitrogen oxide (in kg)
\item \code{so2}: Sulfur dioxide (in kg)
\item \code{co}: Carbon monoxide (in kg)
\item \code{nh3}: Ammonia (in in kg)
\item \code{covnm}: Non-methane volatile organic compounds (in kg)
\item \code{co2}: Carbon dioxide (in ton)
}
}
\seealso{
Other datasets from airparif: \code{\link{emissions_departement}},
  \code{\link{emissions_epci}}, \code{\link{stations}}
}
\concept{datasets from airparif}
\keyword{data}
