% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_magellan_available}
\alias{is_magellan_available}
\title{test airparif availability}
\usage{
is_magellan_available()
}
\description{
check that airparif web services are available.
}
\examples{
is_magellan_available()
}
