% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollutant_label}
\alias{pollutant_label}
\title{Pollutant labels}
\format{An object of class \code{data.frame} with 15 rows and 3 columns.}
\usage{
pollutant_label
}
\description{
Description of pollutants in a data.frame.
\itemize{
\item \code{pollutant}: short code for a pollutant
\item \code{label_fr}: french label of pollutant
\item \code{label_en}: english label of pollutant
}
}
\keyword{data}
