% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emissions.R
\name{read_emissions}
\alias{read_emissions}
\title{air pollutant emissions}
\usage{
read_emissions(geo = "epci")
}
\arguments{
\item{geo}{one of the following values: 'epci', 'departement', 'region'.}
}
\description{
Air pollutant emissions for 2015. Pollutant emissions are the
quantities of pollutants directly released into the atmosphere by human
activities or by natural sources.
}
\examples{
if( is_magellan_available() ){
  read_emissions(geo = "epci")
  read_emissions(geo = "departement")
  read_emissions(geo = "region")
}
}
\seealso{
\code{\link[=is_magellan_available]{is_magellan_available()}}

Other functions about Paris air quality: \code{\link{read_alert}},
  \code{\link{read_atmo}}, \code{\link{read_measures}}
}
\concept{functions about Paris air quality}
