% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{read_measures}
\alias{read_measures}
\title{get pollutants measurement}
\usage{
read_measures(measure = "pm10", granularity = "day",
  dt_start = Sys.Date() - 7, dt_end = Sys.Date(), station_id = NULL,
  record_max = 10000)
}
\arguments{
\item{measure}{one of the following values: 'o3', 'pm10', 'pm25', 'nox', 'no2'.}

\item{granularity}{one of the following values: 'hour', 'day'.}

\item{dt_start, dt_end}{limit dates to use as filters. This parameter
is mandatory.}

\item{station_id}{station_id. If NULL, measures from all stations
are retrieved.}

\item{record_max}{maximum number of records to read}
}
\description{
get measurement of pollutants as a data.frame
}
\note{
The API of airparif is limiting the number of reacords to maximum
2000 records.
}
\examples{
read_measures()
read_measures(measure = "o3", station_id = "4049")
read_measures(measure = "nox",
  dt_start = as.Date("2018-09-01"),
  station_id = "4323")
read_measures(measure = "pm10", station_id = "4181",
  dt_start = as.Date("2019-06-01"), granularity = "hour")
}
