% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations information}
\format{An object of class \code{data.frame} with 21 rows and 11 columns.}
\usage{
stations
}
\description{
Stations information stored in a
data.frame. The dataset contains the following columns:
\itemize{
\item \code{station_ue_code}: european unique code for the station
\item \code{township_name}: township where the station is located
\item \code{township_id}: township identifyer
\item \code{station_name}: station name
\item \code{station_x}: x coordinate the station (lambert 93)
\item \code{station_y}: y coordinate the station (lambert 93)
\item \code{o3}: wether o3 is measured at this station
\item \code{pm10}: wether pm10 is measured at this station
\item \code{pm25}: wether pm25 is measured at this station
\item \code{nox}: wether nox is measured at this station
\item \code{no2}: wether no2 is measured at this station
}
}
\references{
\url{https://data-airparif-asso.opendata.arcgis.com}
}
\author{
Airparif
}
\keyword{data}
