% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atmo.R
\name{read_atmo}
\alias{read_atmo}
\title{ATMO index data}
\usage{
read_atmo()
}
\value{
a data.frame with the following columns:
\itemize{
\item \code{date_time}: \emph{(POSIXct with time zone set to \code{Europe/Paris})}, date-time of index
\item \code{score}: \emph{(integer)}, air quality index
}
}
\description{
get ATMO index data as a data.frame.
The data is only containing the last year of daily historic.
}
\details{
The air quality index ranges from 1 (very good) to 10 (very bad).
It characterizes in a simple and comprehensive way the air quality
of an urban agglomeration.

It is composed of 4 sub index (all ranging from 1 to 10), each being
representative of an air pollutant:
\itemize{
\item fine particles (PM10)
\item ozone (O3)
\item nitrogen dioxide (NO2)
\item sulfur dioxide (SO2)
}

The highest sub-index of the 4 presented above will be the index of the day.

It is not possible to reproduce exact calculations as \code{so2} is not available
through the web service or with \code{\link[=read_measures]{read_measures()}}.
}
\examples{
if( is_magellan_available() ){
  read_atmo()
}
}
\seealso{
Other functions about Paris air quality: \code{\link{read_alert}},
  \code{\link{read_emissions}}, \code{\link{read_measures}}
}
\concept{functions about Paris air quality}
