% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emissions.R
\name{read_emissions}
\alias{read_emissions}
\title{air pollutant emissions}
\usage{
read_emissions(geo = "epci")
}
\arguments{
\item{geo}{one of the following values: "epci", "departement", "region".}
}
\description{
Air pollutant emissions for 2015. Pollutant emissions are the
quantities of pollutants directly released into the atmosphere by human
activities or by natural sources.
}
\examples{
read_emissions(geo = "epci")

read_emissions(geo = "departement")

read_emissions(geo = "region")
}
