% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_measures.R
\name{read_measures}
\alias{read_measures}
\title{get pollutants measurement}
\usage{
read_measures(measure = "pm10", granularity = "day",
  dt_start = Sys.Date() - 7, dt_end = Sys.Date(), station_id = NULL,
  record_max = 2000)
}
\arguments{
\item{measure}{one of the following values: "o3", "pm10", "pm25", "nox", "no2".}

\item{granularity}{one of the following values: "hour", "day".}

\item{dt_start, dt_end}{limit dates to use as filters. This parameter
is mandatory.}

\item{station_id}{station_id. If NULL, measures from all stations
are retrieved. The stations are described in dataset \code{\link{stations}}.}

\item{record_max}{maximum number of records to read}
}
\value{
a data.frame with the following columns:

\code{code_station_ue}: station code where measurements have been carried out.

\code{value}: measurements (unit: microgram per cubic meter)

\code{date}: date of measurement
}
\description{
get measurement of pollutants as a data.frame. Measures
unit is microgram per cubic meter.
}
\note{
The API of airparif is limiting the number of reacords to maximum
2000 records so it will be necessary to create a loop if more rows are
needed.
}
\section{available pollutants}{

The following pollutants are available :

\code{'o3'}: ozone

\code{'pm10'}: Particulate matter with a diameter of 10 micrometers or less

\code{'pm25'}: Particulate matter with a diameter of 2.5 micrometers or less

\code{'nox'}: Nitrogen oxide

\code{'no2'}: Nitrogen dioxide
}

\examples{
read_measures()
read_measures(measure = "o3", station_id = "4049")
read_measures(measure = "nox",
  dt_start = as.Date("2018-09-01"),
  station_id = "4323")
read_measures(measure = "pm10", station_id = "4181",
  dt_start = as.Date("2019-06-01"), granularity = "hour")
}
