% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_measures.R
\name{read_measures}
\alias{read_measures}
\title{get pollutants measurements}
\usage{
read_measures(pollutant = "pm10", granularity = "day",
  dt_start = format(Sys.Date() - 7), dt_end = format(Sys.Date() - 1),
  station_id = NULL, record_max = 2000)
}
\arguments{
\item{pollutant}{one of the following values: 'o3', 'pm10', 'pm25', 'nox', 'no2', 'so2'.}

\item{granularity}{one of the following values: 'hour', 'day'.}

\item{dt_start, dt_end}{limit dates (character formated as `%Y-%m-%d``) to use as filters. This parameter
is mandatory.}

\item{station_id}{station_id. If NULL, measures from all stations
are retrieved. The stations are described in dataset \code{\link{stations}}.}

\item{record_max}{maximum number of records to read}
}
\value{
a data.frame with the following columns:
\itemize{
\item \code{code_station_ue}: \emph{(character)}, station code where measurements have been carried out.
\item \code{value}: \emph{(double)}, measurements (unit: microgram per cubic meter)
\item \code{date_time}: \emph{(POSIXct with time zone set to \code{Europe/Paris})}, date-time of measurement
\item \code{pollutant}: \emph{(character)}, pollutant
}
}
\description{
get measurements of pollutants as a data.frame. All measures
are using unit microgram per cubic meter.
}
\note{
The API of airparif is limiting the number of records to 2000;
it is necessary to loop over dates or stations when more rows are
needed.
}
\section{available pollutants}{


The following pollutants are available :
\itemize{
\item \code{o3}: ozone
\item \code{pm10}: Particulate matter with a diameter of 10 micrometers or less
\item \code{pm25}: Particulate matter with a diameter of 2.5 micrometers or less
\item \code{nox}: Nitrogen oxide
\item \code{no2}: Nitrogen dioxide
}
}

\examples{
if( is_magellan_available() ){

read_measures()
read_measures(pollutant = "pm10", station_id = "4002")
read_measures(pollutant = "pm25", station_id = "4002")
read_measures(pollutant = "o3", station_id = "4049")
read_measures(pollutant = "nox", station_id = "4002")
read_measures(pollutant = "no2", station_id = "4002")
read_measures(pollutant = "nox",
  dt_start = as.Date("2018-09-01"),
  station_id = "4323")
read_measures(pollutant = "pm10", station_id = "4181",
  dt_start = as.Date("2019-06-01"), granularity = "hour")

}
}
\seealso{
\link{stations}, \code{\link[=is_magellan_available]{is_magellan_available()}}

Other functions about Paris air quality: \code{\link{read_alert}},
  \code{\link{read_atmo}}
}
\concept{functions about Paris air quality}
